requirements.txt
discord.py
openai
pylint
pytz
python-dotenv
